test(js_interop_gen): relocate ErrorReporter never-returning method example
Relocate the ErrorReporter test case interface containing a never-returning method to `interfaces_input.d.ts` (and its expected golden `interfaces_expected.dart`) from `functions_input.d.ts`. This addresses Nike's nit by keeping the functions test suite strictly focused on top-level function declarations, moving interface mappings to their dedicated test file.
