+++ js_interop_gen/test/integration/interop_gen/interfaces_expected.dart
@@ -6,6 +6,8 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:js_interop' as _i1;
 
+import 'package:meta/meta.dart' as _i2;
+
 extension type ILogger._(_i1.JSObject _) implements _i1.JSObject {
   external AnonymousUnion_1885558? level;
 
@@ -61,6 +63,10 @@ external Dictionary get dict;
 external LinkedList get rootList;
 @_i1.JS()
 external Comparator<_i1.JSNumber> get compareNumbers;
+extension type ErrorReporter._(_i1.JSObject _) implements _i1.JSObject {
+  @_i2.doNotStore
+  external _i1.JSAny? reportError(String msg);
+}
 extension type const AnonymousUnion_1885558._(String _) {
   static const AnonymousUnion_1885558 debug = AnonymousUnion_1885558._('debug');
